backend/pyproject.toml
[project]
name = "chat-your-notes-backend"
version = "0.1.0"
description = "Flask API for the ChatYourNotes platform"
readme = "../README.md"
requires-python = ">=3.11"
authors = [
    { name = "ChatYourNotes Team" }
]
dependencies = [
    "flask>=3.0,<4.0",
    "flask-cors>=4.0,<5.0",
    "flask-sqlalchemy>=3.1,<4.0",
    "pymysql>=1.1,<2.0",
    "cryptography>=42.0,<45.0",
    "python-dotenv>=1.0,<2.0",
    "pypdf>=4.3,<5.0",
    "google-generativeai>=0.6,<1.0",
    "chromadb>=0.5,<0.6",
    "sentence-transformers>=3.0,<4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "ruff>=0.7,<0.8"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
